Extract edge cases to a separate file
