+++ test/test_unit.py
@@ -1,7 +1,6 @@
 import unittest
 from src import interleave
 from unittest.mock import patch
-import textract
 
 
 class PDFUnitTests(unittest.TestCase):
@@ -29,27 +28,6 @@ class PDFUnitTests(unittest.TestCase):
         with open('text/nbsp_formfeed.txt', 'r') as file:
             cls.nbsp_formfeed = file.read()
 
-        with open('text/edge_line_start_numbers.txt', 'r') as file:
-            cls.edge_numbers = file.read()
-
-        with open('text/edge_missing_paragraphs.txt', 'r') as file:
-            cls.edge_missing_paragraphs = file.read()
-
-        with open('text/edge_first_paragraph.txt', 'r') as file:
-            cls.edge_first_paragraph = file.read()
-
-        with open('text/edge_final_paragraph.txt', 'r') as file:
-            cls.edge_final_paragraph = file.read()
-
-        with open('text/edge_final_paragraph.txt', 'r') as file:
-            cls.edge_final_paragraph = file.read()
-
-        with open('text/edge_EPA_signature_block_1.txt', 'r') as file:
-            cls.edge_EPA_sigblock_1 = file.read()
-
-        with open('text/edge_EPA_signature_block_2.txt', 'r') as file:
-            cls.edge_EPA_sigblock_2 = file.read()
-
         cls.split_simple_text = ['1. First Entry.', '2. Second Entry.', '3. Third Entry.']
 
         cls.split_multiline_text = ['1. First Entry. This is a really long entry. It spans multiple lines. Very long. '
@@ -72,12 +50,6 @@ class PDFUnitTests(unittest.TestCase):
                               ('2. Second Entry.', '2. Second Entry.'),
                               ('3. Third Entry.', '3. Third Entry.')]
 
-        cls.table_title = 'Table 1: PMNs for which EPA untimely published notice of receipt in the Federal Register'
-
-        cls.EPA_signature_1 = '/s/'
-
-        cls.EPA_signature_2 = 'Respectfully submitted,'
-
     def test_opens_PDF(self):
         self.assertEqual(self.short_text, interleave.convert_pdf_to_txt('PDFs/Simple.pdf'))
 
@@ -88,7 +60,7 @@ class PDFUnitTests(unittest.TestCase):
         self.assertEqual(self.multipage_text, interleave.convert_pdf_to_txt('PDFs/Multipage.pdf'))
 
     def test_counts_up_to_1000_paragraphs(self):
-        result = interleave.build_paragraphs(interleave.sanitize_text('\n\n' + self.thousand_paragraphs)) 
+        result = interleave.build_paragraphs(interleave.sanitize_text('\n\n' + self.thousand_paragraphs))
         self.assertEqual(1000, len(result))
 
     def test_builds_paragraphs_correctly(self):
@@ -119,31 +91,6 @@ class PDFUnitTests(unittest.TestCase):
         self.assertNotIn(chr(160), result)
         self.assertNotIn(chr(12), result)
         self.assertEqual(2, len(result))
-
-    def test_edge_case_line_starts_with_numeric_sentence_end(self):
-        result = interleave.build_paragraphs(interleave.sanitize_text('\n\n' + self.edge_numbers))
-        self.assertEqual(7, len(result))  # Expect seven paragraphs
-
-    def test_edge_case_missing_paragraphs(self):
-        result = interleave.build_paragraphs(interleave.sanitize_text('\n\n' + self.edge_missing_paragraphs))
-        self.assertEqual(7, len(result))  # Expect seven paragraphs
-
-    def test_edge_case_isolate_first_paragraph(self):
-        result = interleave.build_paragraphs(interleave.sanitize_text(self.edge_first_paragraph))
-        self.assertEqual(1, len(result))
-
-    def test_edge_case_ignore_trailing_tables(self):
-        result = interleave.sanitize_text(self.edge_final_paragraph)[0]
-        self.assertNotIn(self.table_title, result)
-
-    def test_edge_case_strip_EPA_sigblock_1(self):
-        result = '\n'.join(interleave.sanitize_text(self.edge_EPA_sigblock_1))
-        self.assertNotIn(self.EPA_signature_1, result)
-
-    def test_edge_case_strip_EPA_sigblock_2(self):
-        result = '\n'.join(interleave.sanitize_text(self.edge_EPA_sigblock_2))
-        self.assertNotIn(self.EPA_signature_2, result)
-
 
     def test_zip_sentences_to_tuple(self):
         list1 = '\n\n' + self.short_text
